requirements.txt
Django==4.1.3
django-filter==22.1
djangorestframework==3.14.0
gunicorn==20.1.0
Pillow==9.3.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.24.2
django-rest-passwordreset==1.3.0
django-debug-toolbar==3.7.0
django-taggit==3.1.0